simplify dependency loading
