+++ buildSrc/gradle.properties
@@ -1,1 +1,1 @@
-loom_version=1.14-SNAPSHOT
+loom_version=1.15.5
