+++ common/src/client/java/de/zannagh/armorhider/client/compat/CompatLoader.java
@@ -1,0 +1,5 @@
+package de.zannagh.armorhider.client.compat;
+
+public interface CompatLoader {
+    void init();
+}
