--- common/src/client/java/de/zannagh/armorhider/client/compat/elytratrims/CompatLoader.java
@@ -1,5 +1,0 @@
-package de.zannagh.armorhider.client.compat.elytratrims;
-
-public interface CompatLoader {
-    void init();
-}
