+++ common/versions/fabric-1.20.1/gradle.properties
@@ -1,2 +1,4 @@
 java.version=17
 display_version=mc-1.20.0-1
+elytratrims.version=9K6nhsb4
+geckolib.version=V5QnDeov
