+++ common/versions/fabric-1.21.11/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.11
+elytratrims.version=Nzd1iQCn
+geckolib.version=G1BvHQDL
