+++ common/versions/fabric-1.21.5/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.5-8
+elytratrims.version=Uj37Za5k
+geckolib.version=AyXpbm2h
