+++ common/versions/fabric-1.21.7/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.5-8
+elytratrims.version=QklJxw1S
+geckolib.version=k4Azk0wN
