+++ common/versions/fabric-1.21.9/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.9-10
+elytratrims.version=iLC0LP3D
+geckolib.version=7qjQQSWv
