+++ common/versions/fabric-1.21/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.0-1
+elytratrims.version=KOEvdSw7
+geckolib.version=11VBhLU2
