+++ common/versions/fabric-26.1-rc-2/gradle.properties
@@ -1,2 +1,4 @@
 java.version=25
 display_version=mc-26.1-rc.1-2
+elytratrims.version=NLOF6eKu
+geckolib.version=yUpAp23B
