+++ common/versions/fabric-26.1-snapshot-11/gradle.properties
@@ -1,2 +1,4 @@
 java.version=25
 display_version=26.1-snap.7-11
+elytratrims.version=NLOF6eKu
+geckolib.version=yUpAp23B
