+++ common/versions/fabric-26.1/gradle.properties
@@ -1,2 +1,4 @@
 java.version=25
 display_version=mc-26.1
+elytratrims.version=NLOF6eKu
+geckolib.version=yUpAp23B
