+++ common/versions/neoforge-1.21.1/gradle.properties
@@ -1,0 +1,2 @@
+geckolib.version=UvKZBX05
+elytratrims.version=7mqA6CJO
