+++ common/versions/neoforge-1.21.10/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.9-10
+geckolib.version=xji1VqGU
+elytratrims.version=GxM8lsm4
