+++ common/versions/neoforge-1.21.11/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.11
+geckolib.version=FpONDAt3
+elytratrims.version=u2w6bgX3
