+++ common/versions/neoforge-1.21.4/gradle.properties
@@ -1,2 +1,3 @@
 java.version=21
 display_version=mc-1.21.4
+geckolib.version=eQtABRub
