+++ common/versions/neoforge-1.21.8/gradle.properties
@@ -1,2 +1,4 @@
 java.version=21
 display_version=mc-1.21.5-8
+geckolib.version=KPEgpbCC
+elytratrims.version=RMLnPwtS
