+++ common/versions/neoforge-1.21/gradle.properties
@@ -2,3 +2,5 @@ java.version=21
 neoforge.version=21.0.167
 neoforge.minecraft_version_range=[1.21,1.21.1]
 display_version=mc-1.21.0-1
+geckolib.version=UvKZBX05
+elytratrims.version=7mqA6CJO
