+++ common/versions/neoforge-26.1/gradle.properties
@@ -1,2 +1,4 @@
 java.version=25
 display_version=mc-26.1
+geckolib.version=ScUAwGRA
+elytratrims.version=wgUBgHvL
