+++ fabric/build.gradle.kts
@@ -29,6 +29,9 @@ configure<net.fabricmc.loom.api.LoomGradleExtensionAPI> {
         if (project.isDeobf) {
             vmArg("-Dfabric.gameVersion=${fabricVersion}")
         }
+        if (environment == "client") {
+            programArgs("--username=zannagh")
+        }
     }
 }
