+++ neoforge/build.gradle.kts
@@ -36,40 +36,6 @@ val expandResourcesForIdea = registerExpandResourcesForIdea(
 expandResourcesForIdea.configure { dependsOn(tasks.classes, tasks.named("clientClasses")) }
 patchIdeRunConfigsAllowParallel()
 
-repositories {
-    maven("https://api.modrinth.com/maven") {
-        content { includeGroup("maven.modrinth") }
-    }
-}
-
-// GeckoLib compat — compile-only for verifying mixin targets match.
-val geckoLibNeeded = !project.isDeobf && project.mcVersion.startsWith("1.21.")
-if (geckoLibNeeded) {
-    val mcMinor = project.mcVersion.removePrefix("1.21.").toIntOrNull()
-    dependencies {
-        if (mcMinor != null && mcMinor >= 9) {
-            compileOnly("maven.modrinth:geckolib:xji1VqGU") // GL 5.3-alpha-3 NeoForge 1.21.10
-        } else {
-            compileOnly("maven.modrinth:geckolib:gFmrC8Ru") // GL 4.8.4 NeoForge 1.21.1
-        }
-    }
-}
-
-// ElytraTrims compat — only for versions where compat class compiles (>= 1.21.9)
-// The 1.21.9 artifact uses ResourceLocation which was renamed in 1.21.11, so cap at < 1.21.11.
-val needsElytraTrims = project.isDeobf || project.mcVersion.let {
-    it.startsWith("1.21.") && (it.removePrefix("1.21.").toIntOrNull() ?: 0).let { minor -> minor in 9..10 }
-}
-if (needsElytraTrims) {
-    dependencies {
-        if (project.isDeobf) {
-            compileOnly("maven.modrinth:elytra-trims:fE4eMbIS")       // ET 4.7.0 NeoForge 26.1
-        } else {
-            compileOnly("maven.modrinth:elytra-trims:GxM8lsm4")       // ET 4.5.7 NeoForge 1.21.9+
-        }
-    }
-}
-
 neoForge {
     version = neoforgeVersion
