+++ neoforge/versions/neoforge-1.21.1/gradle.properties
