+++ neoforge/versions/neoforge-1.21.10/gradle.properties
