+++ neoforge/versions/neoforge-1.21.11/gradle.properties
