+++ neoforge/versions/neoforge-1.21.5/gradle.properties
