+++ neoforge/versions/neoforge-1.21.6/gradle.properties
