+++ neoforge/versions/neoforge-1.21.7/gradle.properties
