+++ neoforge/versions/neoforge-1.21.9/gradle.properties
