+++ neoforge/versions/neoforge-26.1.1/gradle.properties
