+++ neoforge/versions/neoforge-26.1/gradle.properties
